+++ integreat_cms/cms/views/language_tree/__init__.py
@@ -2,5 +2,11 @@
 This package contains all views related to the language tree of a region
 """
 from .language_tree_actions import move_language_tree_node, delete_language_tree_node
+from .language_tree_bulk_actions import (
+    BulkMakeVisibleView,
+    BulkHideView,
+    BulkActivateView,
+    BulkDisableView,
+)
 from .language_tree_node_form_view import LanguageTreeNodeCreateView
 from .language_tree_view import LanguageTreeView